Clean up stuff, make volume work again

+++ src/i_sound.h
@@ -118,7 +118,7 @@ void I_UpdateSoundParams(INT32 handle, UINT8 vol, UINT8 sep, UINT8 pitch);
 
 	\return	void
 */
-void I_SetSfxVolume(UINT8 volume);
+void I_SetSfxVolume(int volume);
 
 /// ------------------------
 //  MUSIC SYSTEM
@@ -227,7 +227,7 @@ void I_ResumeSong(void);
 
 	\return	void
 */
-void I_SetMusicVolume(UINT8 volume);
+void I_SetMusicVolume(int volume);
 
 boolean I_SetSongTrack(INT32 track);
